Allow closing modal dialogs by pressing Escape
It's useful to be able to close a modal dialog by pressing escape
instead of using the mouse. This commit fixes an issue with the key
events being handled as desired by preventing keyboard focus on the
top-level modal element. This is consistent with the Bootstrap modal
examples: http://getbootstrap.com/javascript/#modals-examples

Also see:
http://stackoverflow.com/questions/12630156/