Add HTML versions of legal documents
Add HTML versions of the following legal documents provided on Oct 27,
2016:

- ISIC- Medical Disclaimer and Footer (Final).docx
- ISIC- Privacy Policy (Final).docx
- ISIC- Terms of Use (Final).docx

The documents were converted to HTML using Pandoc using a command line
like:

    pandoc -s <filename.docx> -t html -o <filename.html>

This commit contains the full original content of the documents after
being converted. The HTML documents will be hand-modified in future
commits to fix any formatting errors and to remove text that shouldn't
appear in the published documents.