Use camelCase for legal document route fragments
The previous commits served the legal documents from
underscore-separated route fragments, because the provided Terms of Use
refers to a URL containing "privacy_policy". However, that URL as
written wouldn't work anyway, because it doesn't include the hash sign
that indicates a route fragment.

Now, use camelCase for all the new routes, which is consistent with the
rest of the app, and update the URL in the Terms Of Use.