Add a page to request Dataset Contributor access
Logged-in users now always see the "Upload Dataset" button. When
clicked, users not in the "Dataset Contributors" group are shown a page
on which they can request access to the group. When a user requests
access, they are added to the group's pending list and group moderators/
administrators are notified by email.

Currently, the access request view displays placeholder text.

One issue is that there's no notification mechanism to let the user
know when an administrator has acted upon the request. Furthermore,
since the client caches the user's permissions, the user will still have
to manually refresh the page once they are granted permission.

To alleviate the situation a bit, if the user presses the button to
request access again, and the user is now part of the group, the client
permissions will update and the page will refresh to show the upload
dataset view.