Add alert dialog with a single button
The dialog provided by girder.confirm() is inappropriate outside of the
"OK/Cancel" use case. For example, when a dialog's intent is to alert a
user and the action can't be cancelled, the dialog should display only
one button and should provide a callback for the button or when the
dialog closes another way (e.g. the user presses Escape).

This alert dialog supports this use case, while the dialog provided by
girder.confirm() does not.