Send email confirmation when dataset is uploaded
Send an email confirmation once a dataset is uploaded. The email
includes all the dataset information that the user supplied on the
upload page, as well as a link to the agreed upon Terms of Use document.

Additionally, send an email to the 'Dataset QC Reviewers' group to
notify them about the new dataset. This email also includes the dataset
information.