Improve upload widget drop zone style
This commit improves the style applied to the upload widget, especially
with regards to drag and drop.

The drop zone is taller and wider. This makes it easier to drag and drop
to upload a file.

Additionally, the width of "Browse or drop files" button no longer
depends on the width of the text below (which changes after uploading a
file).