Replace ApplyMetadataValidationView with partial render of the top view
Now that ApplyMetadataValidationView triggers no events and simply
updates in response to model/collection changes, it makes sense to
remove the view and instead render the validation section directly in
the top view.

As a result, this removes the need to copy the MetadataFileModel
attributes, which was used only as a mechanism to trigger the validation
view update.