Don't attempt to send email to an empty list of addresses
When a user requests permission to create a dataset the moderators and
administrators of the 'Dataset Contributors' group receive an email.
When the group has no moderators or administrators, such as on a newly
provisioned deployment, the sendEmail() function throws an exception.

To avoid the exception, add a check that the list of email addresses is
not empty before attempting to send email.