Make single image upload a two-step process
Uploading a single image to a dataset is now a two-step process.

In the first step, metadata such as the image filename and signature are
sent and a contentId is returned.

In the second step, the contentId is used to form a URL to which to send
the raw image data. The image data is sent in the request body. The
endpoint implementation notes, visible on Swagger UI, include JavaScript
examples of sending binary data from a file in a request body.