Add study participation request API
A user can request to participate in a study using the following
endpoint:

    POST /study/{id}/participate

Study administrators receive an email notification of the request. A
list of pending requests is returned to study administrators in a
study's "waitingUsers" field.

Study administrators can accept a request by adding the user to the
study with:

    POST /study/{id}/users

Adding the user to the study removes any study participation request
from the user.

Study administrators can delete the request with the following endpoint:

    DELETE /study/{id}/participate/{userId}