Update metadata API
Update the API to register, apply, and download metadata to avoid using
Girder folder/file endpoints.

The `POST /dataset/{id}/metadata` endpoint now takes the metadata CSV
data in the HTTP request body and the filename as a query parameter.

The `POST /dataset/{id}/metadata/{metadataFileId}` endpoint is unchanged
except that the 'fileId' parameter is renamed to 'metadataFileId'.

The `GET /dataset/{id}/metadata/{metadataFileId}` endpoint is added to
download a metadata file.