Add API for clients to upload ZIP files of images directly to S3
Add endpoints to the dataset resource to support clients uploading ZIP
files of images directly to S3. Once the file is uploaded, calling the
existing "POST /dataset/{id}/imageZip" endpoint adds the images in the
ZIP file to a dataset.

To start a new direct-to-S3 upload of a ZIP file of images, call "POST
/dataset/{id}/zip". Depending on the size of the ZIP file, one of two
workflows is followed to upload the file to S3: single-part upload or
multi-part upload.

In general, each HTTP request to these endpoints responds with
information for the client to make a subsequent HTTP request to S3. The
returned information includes the HTTP method, a pre-signed URL, and
headers. The client is responsible to add headers such as
Content-Length, Content-MD5, and additional data, as appropriate.

In the single-part upload workflow, the client uploads the file data to
S3 in a single request. The client then notifies the server that the
upload is complete by calling "POST
/dataset/{id}/zip/{zipUploadId}/completion".

In the multi-part upload workflow, the client uploads the file data to
S3 in multiple requests. For each part of the file, the client calls
"POST /dataset/{id}/zip/{part}" to get the S3 request information and
then makes the S3 request, including the data for the part. Once all the
parts have been uploaded, the client notifies the server that the upload
is complete by calling "POST /dataset/{id}/zip/{zipUploadId}/completion".
The response to this request includes information to make an additional
S3 request to finalize the upload.

See the documentation for each endpoint for details. Additionally, see
the AWS S3 Uploading Objects documentation for supplemental information
on the S3 API:
https://docs.aws.amazon.com/AmazonS3/latest/dev/UploadingObjects.html