Load attached metadata file using same access level required for dataset
Because the metadata files are attached to datasets, loading them with a
required access level actually checks the permissions on the dataset.
Therefore, use the same access level for loading the metadata files as
when loading the dataset.

Note that this implementation results in redundant permission checking.
An alternative would be to force load the files, but the current
strategy is safer in case the file storage convention changes in the
future.