Enable direct-to-S3 ZIP upload using temporary security credentials
Add endpoints that enable clients to upload ZIP files of images directly
to S3. The client receives temporary security credentials for AWS and
can upload the file using an AWS SDK. This workflow enables faster
transfers, because the server is not involved in the upload.
Furthermore, by delegating the upload step to the AWS API, the image
ingestion process becomes simpler and easier to implement in a new
client app.